misc testing utility improvements
